Remove Blueprint container for the romaji tag. Should re-design it if needed.
