+++ osu.Game.Rulesets.Karaoke/Screens/Edit/Beatmaps/Lyrics/Components/Lyrics/Blueprints/RubyTagSelectionBlueprint.cs
@@ -8,21 +8,28 @@ using osu.Framework.Extensions;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
 using osu.Framework.Graphics.Cursor;
+using osu.Framework.Graphics.Primitives;
+using osu.Framework.Graphics.Shapes;
 using osu.Framework.Graphics.UserInterface;
 using osu.Framework.Input.Events;
 using osu.Game.Graphics;
 using osu.Game.Graphics.UserInterfaceV2;
+using osu.Game.Rulesets.Edit;
 using osu.Game.Rulesets.Karaoke.Edit.ChangeHandlers.Lyrics;
 using osu.Game.Rulesets.Karaoke.Objects;
 using osu.Game.Rulesets.Karaoke.Screens.Edit.Beatmaps.Lyrics.States.Modes;
 using osuTK;
+using osuTK.Graphics;
 
 namespace osu.Game.Rulesets.Karaoke.Screens.Edit.Beatmaps.Lyrics.Components.Lyrics.Blueprints;
 
-public partial class RubyTagSelectionBlueprint : TextTagSelectionBlueprint<RubyTag>, IHasPopover
+public partial class RubyTagSelectionBlueprint : SelectionBlueprint<RubyTag>, IHasPopover
 {
     [Resolved]
     private IEditRubyModeState editRubyModeState { get; set; } = null!;
+
+    [Resolved]
+    private IPreviewLyricPositionProvider previewLyricPositionProvider { get; set; } = null!;
 
     [UsedImplicitly]
     private readonly Bindable<string> text;
@@ -32,6 +39,9 @@ public partial class RubyTagSelectionBlueprint : TextTagSelectionBlueprint<RubyT
 
     [UsedImplicitly]
     private readonly BindableNumber<int> endIndex;
+
+    private readonly Container previewTextArea;
+    private readonly Container indexRangeBackground;
 
     public RubyTagSelectionBlueprint(RubyTag item)
         : base(item)
@@ -43,15 +53,91 @@ public partial class RubyTagSelectionBlueprint : TextTagSelectionBlueprint<RubyT
         text = item.TextBindable.GetBoundCopy();
         startIndex = item.StartIndexBindable.GetBoundCopy();
         endIndex = item.EndIndexBindable.GetBoundCopy();
+
+        InternalChildren = new[]
+        {
+            previewTextArea = new Container
+            {
+                Alpha = 0,
+            },
+            indexRangeBackground = new Container
+            {
+                Masking = true,
+                BorderThickness = 3,
+                Alpha = 0,
+                BorderColour = Color4.White,
+                Children = new Drawable[]
+                {
+                    new Box
+                    {
+                        RelativeSizeAxes = Axes.Both,
+                        Alpha = 0f,
+                        AlwaysPresent = true,
+                    },
+                },
+            },
+        };
     }
 
     [BackgroundDependencyLoader]
-    private void load()
+    private void load(OsuColour colours)
     {
+        indexRangeBackground.Colour = colours.Pink;
+
         UpdatePositionAndSize();
         text.BindValueChanged(_ => UpdatePositionAndSize());
         startIndex.BindValueChanged(_ => UpdatePositionAndSize());
         endIndex.BindValueChanged(_ => UpdatePositionAndSize());
+    }
+
+    protected void UpdatePositionAndSize()
+    {
+        // wait until lyric update ruby position.
+        ScheduleAfterChildren(() =>
+        {
+            var rubyTagRect = previewLyricPositionProvider.GetTextTagByPosition(Item);
+
+            if (rubyTagRect == null)
+            {
+                return;
+            }
+
+            var startRect = previewLyricPositionProvider.GetRectByCharIndex(Item.StartIndex);
+            var endRect = previewLyricPositionProvider.GetRectByCharIndex(Item.EndIndex);
+
+            // update select position
+            updateDrawableRect(previewTextArea, rubyTagRect.Value);
+
+            // update index range position.
+            var indexRangePosition = new Vector2(startRect.Left, rubyTagRect.Value.Y);
+            var indexRangeSize = new Vector2(endRect.Right - startRect.Left, rubyTagRect.Value.Height);
+            updateDrawableRect(indexRangeBackground, new RectangleF(indexRangePosition, indexRangeSize));
+        });
+
+        static void updateDrawableRect(Drawable target, RectangleF rect)
+        {
+            target.X = rect.X;
+            target.Y = rect.Y;
+            target.Width = rect.Width;
+            target.Height = rect.Height;
+        }
+    }
+
+    public override bool ReceivePositionalInputAt(Vector2 screenSpacePos)
+        => previewTextArea.ReceivePositionalInputAt(screenSpacePos);
+
+    public override Vector2 ScreenSpaceSelectionPoint => ScreenSpaceDrawQuad.TopLeft;
+
+    public override Quad SelectionQuad => previewTextArea.ScreenSpaceDrawQuad;
+
+    protected override void OnSelected()
+    {
+        indexRangeBackground.FadeIn(500);
+    }
+
+    protected override void OnDeselected()
+    {
+        indexRangeBackground.FadeOut(500);
     }
 
     public Popover GetPopover() => new RubyEditPopover(Item);
